+++ nap/rest/views/object.py
@@ -29,7 +29,7 @@ class ObjectPutMixin:
         try:
             self.mapper._apply(self.data)
         except ValidationError as e:
-            return self.put_invalid(e.error_dict)
+            return self.put_invalid(self.mapper._errors)
 
         return self.put_valid()
 
@@ -52,7 +52,7 @@ class ObjectPatchMixin:
         try:
             self.mapper._patch(self.data)
         except ValidationError as e:
-            return self.patch_invalid(e.error_dict)
+            return self.patch_invalid(self.mapper._errors)
 
         return self.patch_valid()